staketia migration - redeem stake (#1215)
