+++ x/stakeibc/keeper/host_zone.go
@@ -288,3 +288,16 @@ func (k Keeper) GetTargetValAmtsForHostZone(ctx sdk.Context, hostZone types.Host
 
 	return targetUnbondingsByValidator, nil
 }
+
+// Enables redemptions by setting the parameter on the host zone to true
+// This is used during the staketia/stakedym migrations
+func (k Keeper) EnableRedemptions(ctx sdk.Context, chainId string) error {
+	hostZone, found := k.GetHostZone(ctx, chainId)
+	if !found {
+		return types.ErrHostZoneNotFound.Wrapf(chainId)
+	}
+
+	hostZone.RedemptionsEnabled = true
+	k.SetHostZone(ctx, hostZone)
+	return nil
+}
