+++ x/stakeibc/types/errors.go
@@ -67,4 +67,5 @@ var (
 	ErrValidatorExceedsWeightCap           = errorsmod.Register(ModuleName, 1560, "validator exceeds weight cap")
 	ErrFeeSplitInvariantFailed             = errorsmod.Register(ModuleName, 1561, "failed to calculate fee split")
 	ErrFailedToRegisterRebate              = errorsmod.Register(ModuleName, 1562, "failed to register rebate")
+	ErrRedemptionsDisabled                 = errorsmod.Register(ModuleName, 1563, "redemptions disabled")
 )
