+++ x/staketia/keeper/unbonding_test.go
@@ -9,6 +9,8 @@ import (
 
 	"github.com/Stride-Labs/stride/v22/app/apptesting"
 	"github.com/Stride-Labs/stride/v22/utils"
+	epochtypes "github.com/Stride-Labs/stride/v22/x/epochs/types"
+	recordtypes "github.com/Stride-Labs/stride/v22/x/records/types"
 	stakeibctypes "github.com/Stride-Labs/stride/v22/x/stakeibc/types"
 	"github.com/Stride-Labs/stride/v22/x/staketia/types"
 )
@@ -30,19 +32,22 @@ type RedeemStakeTestCase struct {
 
 	userAccount      Account
 	hostZone         *types.HostZone
+	stakeibcHostZone *stakeibctypes.HostZone
 	accUnbondRecord  *types.UnbondingRecord
 	redemptionRecord *types.RedemptionRecord
 	redeemMsg        types.MsgRedeemStake
 
-	expectedUnbondingRecord  *types.UnbondingRecord
-	expectedRedemptionRecord *types.RedemptionRecord
-	expectedErrorContains    string
+	expectedUnbondingRecord          *types.UnbondingRecord
+	expectedRedemptionRecord         *types.RedemptionRecord
+	expectedStakeibcRedemptionAmount sdkmath.Int
+	expectedErrorContains            string
 }
 
 // Create the correct amounts in accounts, setup the records in store
 func (s *KeeperTestSuite) SetupTestRedeemStake(
 	userAccount Account,
 	hostZone *types.HostZone,
+	stakeibcHostZone *stakeibctypes.HostZone,
 	accUnbondRecord *types.UnbondingRecord,
 	redemptionRecord *types.RedemptionRecord,
 ) {
@@ -51,6 +56,10 @@ func (s *KeeperTestSuite) SetupTestRedeemStake(
 
 	if hostZone != nil {
 		s.App.StaketiaKeeper.SetHostZone(s.Ctx, *hostZone)
+	}
+
+	if stakeibcHostZone != nil {
+		s.App.StakeibcKeeper.SetHostZone(s.Ctx, *stakeibcHostZone)
 	}
 
 	if accUnbondRecord != nil {
@@ -69,18 +78,39 @@ func (s *KeeperTestSuite) SetupTestRedeemStake(
 	if redemptionRecord != nil {
 		s.App.StaketiaKeeper.SetRedemptionRecord(s.Ctx, *redemptionRecord)
 	}
+
+	// Prepare stakeibc for a redemption
+	epochNumber := uint64(1)
+	s.App.StakeibcKeeper.SetEpochTracker(s.Ctx, stakeibctypes.EpochTracker{
+		EpochIdentifier: epochtypes.DAY_EPOCH,
+		EpochNumber:     epochNumber,
+	})
+
+	s.App.RecordsKeeper.SetEpochUnbondingRecord(s.Ctx, recordtypes.EpochUnbondingRecord{
+		EpochNumber: epochNumber,
+		HostZoneUnbondings: []*recordtypes.HostZoneUnbonding{
+			{
+				HostZoneId:        types.CelestiaChainId,
+				StTokenAmount:     sdkmath.ZeroInt(),
+				NativeTokenAmount: sdkmath.ZeroInt(),
+			},
+		},
+	})
 }
 
 // Default values for key variables, different tests will change 1-2 fields for setup
 func (s *KeeperTestSuite) getDefaultTestInputs() (
 	*Account,
 	*types.HostZone,
+	*stakeibctypes.HostZone,
 	*types.UnbondingRecord,
 	*types.RedemptionRecord,
 	*types.MsgRedeemStake,
 ) {
 	redeemerAccount := s.TestAccs[0]
 	redemptionAccount := s.TestAccs[1]
+	receiverAddress := s.TestAccs[2].String()
+	depositAddress := s.TestAccs[3].String()
 
 	defaultUserAccount := Account{
 		account:      redeemerAccount,
@@ -89,16 +119,23 @@ func (s *KeeperTestSuite) getDefaultTestInputs() (
 	}
 
 	redemptionRate := sdk.MustNewDecFromStr("1.1")
-	defaultHostZone := types.HostZone{
-		NativeTokenDenom:       HostNativeDenom,
-		RedemptionAddress:      redemptionAccount.String(),
+	defaultStaketiaHostZone := types.HostZone{
+		NativeTokenDenom:          HostNativeDenom,
+		RedemptionAddress:         redemptionAccount.String(),
+		RemainingDelegatedBalance: sdkmath.NewInt(1_000_000_000),
+		Halted:                    false,
+	}
+	defaultStakeibcHostZone := stakeibctypes.HostZone{
+		ChainId:                types.CelestiaChainId,
+		HostDenom:              HostNativeDenom,
+		Bech32Prefix:           "stride",
+		DepositAddress:         depositAddress,
 		RedemptionRate:         redemptionRate,
 		MinRedemptionRate:      redemptionRate.Sub(sdk.MustNewDecFromStr("0.2")),
 		MinInnerRedemptionRate: redemptionRate.Sub(sdk.MustNewDecFromStr("0.1")),
 		MaxInnerRedemptionRate: redemptionRate.Add(sdk.MustNewDecFromStr("0.1")),
 		MaxRedemptionRate:      redemptionRate.Add(sdk.MustNewDecFromStr("0.2")),
-		DelegatedBalance:       sdkmath.NewInt(1_000_000_000),
-		Halted:                 false,
+		TotalDelegations:       sdkmath.NewInt(1_000_000_000),
 	}
 
 	defaultAccUnbondingRecord := types.UnbondingRecord{
@@ -120,21 +157,32 @@ func (s *KeeperTestSuite) getDefaultTestInputs() (
 	defaultMsg := types.MsgRedeemStake{
 		Redeemer:      redeemerAccount.String(),
 		StTokenAmount: sdk.NewInt(1_000_000),
+		Receiver:      receiverAddress,
 	}
 
-	return &defaultUserAccount, &defaultHostZone, &defaultAccUnbondingRecord,
-		&defaultRedemptionRecord, &defaultMsg
+	return &defaultUserAccount, &defaultStaketiaHostZone, &defaultStakeibcHostZone,
+		&defaultAccUnbondingRecord, &defaultRedemptionRecord, &defaultMsg
 }
 
 func (s *KeeperTestSuite) TestRedeemStake() {
-	defaultUA, defaultHZ, defaultUR, defaultRR, defaultMsg := s.getDefaultTestInputs()
+	defaultUA, defaultMsHZ, defaultIcaHZ, defaultUR, defaultRR, defaultMsg := s.getDefaultTestInputs()
 
 	testCases := []RedeemStakeTestCase{
 		{
 			testName: "[Error] Can't find the HostZone",
 
-			userAccount: *defaultUA,
-			hostZone:    nil,
+			userAccount:      *defaultUA,
+			hostZone:         nil,
+			stakeibcHostZone: defaultIcaHZ,
+
+			expectedErrorContains: types.ErrHostZoneNotFound.Error(),
+		},
+		{
+			testName: "[Error] Can't find the stakeibc HostZone",
+
+			userAccount:      *defaultUA,
+			hostZone:         defaultMsHZ,
+			stakeibcHostZone: nil,
 
 			expectedErrorContains: types.ErrHostZoneNotFound.Error(),
 		},
@@ -143,10 +191,11 @@ func (s *KeeperTestSuite) TestRedeemStake() {
 
 			userAccount: *defaultUA,
 			hostZone: func() *types.HostZone {
-				_, hz, _, _, _ := s.getDefaultTestInputs()
+				_, hz, _, _, _, _ := s.getDefaultTestInputs()
 				hz.RedemptionAddress = "nonparsable-address"
 				return hz
 			}(),
+			stakeibcHostZone: defaultIcaHZ,
 
 			expectedErrorContains: "could not bech32 decode redemption address",
 		},
@@ -155,10 +204,11 @@ func (s *KeeperTestSuite) TestRedeemStake() {
 
 			userAccount: *defaultUA,
 			hostZone: func() *types.HostZone {
-				_, hz, _, _, _ := s.getDefaultTestInputs()
+				_, hz, _, _, _, _ := s.getDefaultTestInputs()
 				hz.Halted = true
 				return hz
 			}(),
+			stakeibcHostZone: defaultIcaHZ,
 
 			expectedErrorContains: types.ErrHostZoneHalted.Error(),
 		},
@@ -166,8 +216,9 @@ func (s *KeeperTestSuite) TestRedeemStake() {
 			testName: "[Error] RedemptionRate outside of bounds",
 
 			userAccount: *defaultUA,
-			hostZone: func() *types.HostZone {
-				_, hz, _, _, _ := s.getDefaultTestInputs()
+			hostZone:    defaultMsHZ,
+			stakeibcHostZone: func() *stakeibctypes.HostZone {
+				_, _, hz, _, _, _ := s.getDefaultTestInputs()
 				hz.RedemptionRate = sdk.MustNewDecFromStr("5.2")
 				return hz
 			}(),
@@ -177,9 +228,10 @@ func (s *KeeperTestSuite) TestRedeemStake() {
 		{
 			testName: "[Error] No Accumulating UndondingRecord",
 
-			userAccount:     *defaultUA,
-			hostZone:        defaultHZ,
-			accUnbondRecord: nil,
+			userAccount:      *defaultUA,
+			hostZone:         defaultMsHZ,
+			stakeibcHostZone: defaultIcaHZ,
+			accUnbondRecord:  nil,
 
 			expectedErrorContains: types.ErrBrokenUnbondingRecordInvariant.Error(),
 		},
@@ -187,13 +239,14 @@ func (s *KeeperTestSuite) TestRedeemStake() {
 			testName: "[Error] Not enough tokens in wallet",
 
 			userAccount: func() Account {
-				acc, _, _, _, _ := s.getDefaultTestInputs()
+				acc, _, _, _, _, _ := s.getDefaultTestInputs()
 				acc.stTokens.Amount = sdk.NewInt(500_000)
 				return *acc
 			}(),
-			hostZone:        defaultHZ,
-			accUnbondRecord: defaultUR,
-			redeemMsg:       *defaultMsg, // attempt to redeem 1_000_000 stTokens
+			hostZone:         defaultMsHZ,
+			stakeibcHostZone: defaultIcaHZ,
+			accUnbondRecord:  defaultUR,
+			redeemMsg:        *defaultMsg, // attempt to redeem 1_000_000 stTokens
 
 			expectedErrorContains: sdkerrors.ErrInsufficientFunds.Error(),
 		},
@@ -201,14 +254,15 @@ func (s *KeeperTestSuite) TestRedeemStake() {
 			testName: "[Error] Redeeming more than HostZone delegation total",
 
 			userAccount: func() Account {
-				acc, _, _, _, _ := s.getDefaultTestInputs()
+				acc, _, _, _, _, _ := s.getDefaultTestInputs()
 				acc.stTokens.Amount = sdk.NewInt(5_000_000_000)
 				return *acc
 			}(),
-			hostZone:        defaultHZ, // 1_000_000_000 total delegation
-			accUnbondRecord: defaultUR,
+			hostZone:         defaultMsHZ, // 1_000_000_000 total delegation
+			stakeibcHostZone: defaultIcaHZ,
+			accUnbondRecord:  defaultUR,
 			redeemMsg: func() types.MsgRedeemStake {
-				_, _, _, _, msg := s.getDefaultTestInputs()
+				_, _, _, _, _, msg := s.getDefaultTestInputs()
 				msg.StTokenAmount = sdk.NewInt(5_000_000_000)
 				return *msg
 			}(),
@@ -216,16 +270,56 @@ func (s *KeeperTestSuite) TestRedeemStake() {
 			expectedErrorContains: types.ErrUnbondAmountToLarge.Error(),
 		},
 		{
+			testName: "[Error] Redeeming is disabled",
+
+			userAccount: *defaultUA,
+			hostZone: func() *types.HostZone {
+				_, hz, _, _, _, _ := s.getDefaultTestInputs()
+				hz.RemainingDelegatedBalance = sdkmath.ZeroInt()
+				return hz
+			}(),
+			stakeibcHostZone: defaultIcaHZ,
+			accUnbondRecord:  defaultUR,
+			redeemMsg:        *defaultMsg,
+
+			expectedErrorContains: types.ErrRedemptionsDisabled.Error(),
+		},
+		{
+			testName: "[Error] Spillover fails in stakeibc",
+
+			userAccount: *defaultUA,
+			hostZone: func() *types.HostZone {
+				_, hz, _, _, _, msg := s.getDefaultTestInputs()
+				hz.RemainingDelegatedBalance = msg.StTokenAmount.Sub(sdkmath.OneInt()) // subtract 1 so there's excess
+				return hz
+			}(),
+			stakeibcHostZone: func() *stakeibctypes.HostZone {
+				_, _, hz, _, _, _ := s.getDefaultTestInputs()
+				hz.RedemptionRate = sdk.OneDec()
+				return hz
+			}(),
+			accUnbondRecord: defaultUR,
+			redeemMsg: func() types.MsgRedeemStake {
+				// Make receiver address invalid so stakeibc fails
+				_, _, _, _, _, msg := s.getDefaultTestInputs()
+				msg.Receiver = ""
+				return *msg
+			}(),
+
+			expectedErrorContains: "unable to execute stakeibc redeem stake",
+		},
+		{
 			testName: "[Success] No RR exists yet, RedeemStake tx creates one",
 
 			userAccount:      *defaultUA,
-			hostZone:         defaultHZ,
+			hostZone:         defaultMsHZ,
+			stakeibcHostZone: defaultIcaHZ,
 			accUnbondRecord:  defaultUR,
 			redemptionRecord: nil,
 			redeemMsg:        *defaultMsg, // redeem 1_000_000 stTokens
 
 			expectedUnbondingRecord: func() *types.UnbondingRecord {
-				_, hz, ur, _, msg := s.getDefaultTestInputs()
+				_, _, hz, ur, _, msg := s.getDefaultTestInputs()
 				ur.StTokenAmount = ur.StTokenAmount.Add(msg.StTokenAmount)
 				nativeDiff := sdk.NewDecFromInt(msg.StTokenAmount).Mul(hz.RedemptionRate).TruncateInt()
 				ur.NativeAmount = ur.NativeAmount.Add(nativeDiff)
@@ -235,32 +329,92 @@ func (s *KeeperTestSuite) TestRedeemStake() {
 				UnbondingRecordId: defaultUR.Id,
 				Redeemer:          defaultMsg.Redeemer,
 				StTokenAmount:     defaultMsg.StTokenAmount,
-				NativeAmount:      sdk.NewDecFromInt(defaultMsg.StTokenAmount).Mul(defaultHZ.RedemptionRate).TruncateInt(),
+				NativeAmount:      sdk.NewDecFromInt(defaultMsg.StTokenAmount).Mul(defaultIcaHZ.RedemptionRate).TruncateInt(),
 			},
 		},
 		{
 			testName: "[Success] RR exists already for redeemer, RedeemStake tx updates",
 
 			userAccount:      *defaultUA,
-			hostZone:         defaultHZ,
+			hostZone:         defaultMsHZ,
+			stakeibcHostZone: defaultIcaHZ,
 			accUnbondRecord:  defaultUR,
 			redemptionRecord: defaultRR,   // previous redeemption of 400_000
 			redeemMsg:        *defaultMsg, // redeem 1_000_000 stTokens
 
 			expectedUnbondingRecord: func() *types.UnbondingRecord {
-				_, hz, ur, _, msg := s.getDefaultTestInputs()
+				_, _, hz, ur, _, msg := s.getDefaultTestInputs()
 				ur.StTokenAmount = ur.StTokenAmount.Add(msg.StTokenAmount)
 				nativeDiff := sdk.NewDecFromInt(msg.StTokenAmount).Mul(hz.RedemptionRate).TruncateInt()
 				ur.NativeAmount = ur.NativeAmount.Add(nativeDiff)
 				return ur
 			}(),
 			expectedRedemptionRecord: func() *types.RedemptionRecord {
-				_, hz, _, rr, msg := s.getDefaultTestInputs()
+				_, _, hz, _, rr, msg := s.getDefaultTestInputs()
 				rr.StTokenAmount = rr.StTokenAmount.Add(msg.StTokenAmount)
 				nativeDiff := sdk.NewDecFromInt(msg.StTokenAmount).Mul(hz.RedemptionRate).TruncateInt()
 				rr.NativeAmount = rr.NativeAmount.Add(nativeDiff)
 				return rr
 			}(),
+		},
+		{
+			testName: "[Success] Redeems all remaining balance",
+
+			userAccount: *defaultUA,
+			hostZone: func() *types.HostZone {
+				_, msHz, icaHz, _, _, msg := s.getDefaultTestInputs()
+				nativeRedeemAmount := sdk.NewDecFromInt(msg.StTokenAmount).Mul(icaHz.RedemptionRate).TruncateInt()
+				msHz.RemainingDelegatedBalance = nativeRedeemAmount
+				return msHz
+			}(),
+			stakeibcHostZone: defaultIcaHZ,
+			accUnbondRecord:  defaultUR,
+			redeemMsg:        *defaultMsg,
+
+			expectedUnbondingRecord: func() *types.UnbondingRecord {
+				_, _, hz, ur, _, msg := s.getDefaultTestInputs()
+				ur.StTokenAmount = ur.StTokenAmount.Add(msg.StTokenAmount)
+				nativeDiff := sdk.NewDecFromInt(msg.StTokenAmount).Mul(hz.RedemptionRate).TruncateInt()
+				ur.NativeAmount = ur.NativeAmount.Add(nativeDiff)
+				return ur
+			}(),
+			expectedRedemptionRecord: &types.RedemptionRecord{
+				UnbondingRecordId: defaultUR.Id,
+				Redeemer:          defaultMsg.Redeemer,
+				StTokenAmount:     defaultMsg.StTokenAmount,
+				NativeAmount:      sdk.NewDecFromInt(defaultMsg.StTokenAmount).Mul(defaultIcaHZ.RedemptionRate).TruncateInt(),
+			},
+		},
+		{
+			testName: "[Success] Redeems with stakeibc spillover",
+
+			userAccount: *defaultUA,
+			hostZone: func() *types.HostZone {
+				_, hz, _, _, _, msg := s.getDefaultTestInputs()
+				hz.RemainingDelegatedBalance = msg.StTokenAmount.Sub(sdkmath.OneInt()) // subtract 1 so there's excess
+				return hz
+			}(),
+			stakeibcHostZone: func() *stakeibctypes.HostZone {
+				_, _, hz, _, _, _ := s.getDefaultTestInputs()
+				hz.RedemptionRate = sdk.OneDec()
+				return hz
+			}(),
+			accUnbondRecord: defaultUR,
+			redeemMsg:       *defaultMsg,
+
+			expectedUnbondingRecord: func() *types.UnbondingRecord {
+				_, _, _, ur, _, msg := s.getDefaultTestInputs()
+				ur.StTokenAmount = ur.StTokenAmount.Add(msg.StTokenAmount).Sub(sdkmath.OneInt())
+				ur.NativeAmount = ur.NativeAmount.Add(msg.StTokenAmount).Sub(sdkmath.OneInt())
+				return ur
+			}(),
+			expectedRedemptionRecord: &types.RedemptionRecord{
+				UnbondingRecordId: defaultUR.Id,
+				Redeemer:          defaultMsg.Redeemer,
+				StTokenAmount:     defaultMsg.StTokenAmount.Sub(sdkmath.OneInt()),
+				NativeAmount:      defaultMsg.StTokenAmount.Sub(sdkmath.OneInt()),
+			},
+			expectedStakeibcRedemptionAmount: sdkmath.OneInt(),
 		},
 	}
 
@@ -274,7 +428,7 @@ func (s *KeeperTestSuite) TestRedeemStake() {
 
 func (s *KeeperTestSuite) checkRedeemStakeTestCase(tc RedeemStakeTestCase) {
 	s.SetupTest() // reset state
-	s.SetupTestRedeemStake(tc.userAccount, tc.hostZone, tc.accUnbondRecord, tc.redemptionRecord)
+	s.SetupTestRedeemStake(tc.userAccount, tc.hostZone, tc.stakeibcHostZone, tc.accUnbondRecord, tc.redemptionRecord)
 
 	startingStEscrowBalance := sdk.NewInt64Coin(StDenom, 0)
 	if tc.hostZone != nil {
@@ -285,7 +439,7 @@ func (s *KeeperTestSuite) checkRedeemStakeTestCase(tc RedeemStakeTestCase) {
 	}
 
 	// Run the RedeemStake, verify expected errors returned or no errors with expected updates to records
-	_, err := s.App.StaketiaKeeper.RedeemStake(s.Ctx, tc.redeemMsg.Redeemer, tc.redeemMsg.StTokenAmount)
+	_, err := s.App.StaketiaKeeper.RedeemStake(s.Ctx, tc.redeemMsg.Redeemer, tc.redeemMsg.Receiver, tc.redeemMsg.StTokenAmount)
 	if tc.expectedErrorContains == "" {
 		// Successful Run Test Case
 		s.Require().NoError(err, "No error expected during redeem stake execution")
@@ -307,6 +461,14 @@ func (s *KeeperTestSuite) checkRedeemStakeTestCase(tc RedeemStakeTestCase) {
 		currentStEscrowBalance := s.App.BankKeeper.GetBalance(s.Ctx, escrowAccount, StDenom)
 		s.Require().NotEqual(startingStEscrowBalance, currentStEscrowBalance, "Escrowed balance should have changed")
 		s.Require().Equal(currentStEscrowBalance.Amount, currentAUR.StTokenAmount, "Escrowed balance does not match the UnbondingRecord")
+
+		// If there's stakeibc spillover, check that the amount was stored in the user redemption record
+		if !tc.expectedStakeibcRedemptionAmount.IsNil() {
+			userRedemptionRecords := s.App.RecordsKeeper.GetAllUserRedemptionRecord(s.Ctx)
+			s.Require().Len(userRedemptionRecords, 1, "there should be a stakeibc user redemption record")
+			s.Require().Equal(tc.expectedStakeibcRedemptionAmount.Int64(), userRedemptionRecords[0].StTokenAmount.Int64(),
+				"stakeibc user redemption record amount")
+		}
 	} else {
 		// Expected Error Test Case
 		s.Require().Error(err, "Error expected to be returned but none found")
@@ -333,7 +495,7 @@ func (s *KeeperTestSuite) TestPrepareUndelegation() {
 	oldRedemptionRate := sdk.MustNewDecFromStr("1.9")
 	redemptionRate := sdk.MustNewDecFromStr("1.999")
 	s.App.StakeibcKeeper.SetHostZone(s.Ctx, stakeibctypes.HostZone{
-		ChainId:        HostChainId,
+		ChainId:        types.CelestiaChainId,
 		RedemptionRate: redemptionRate,
 	})
 
@@ -420,12 +582,13 @@ func (s *KeeperTestSuite) TestPrepareUndelegation() {
 	s.Require().ErrorContains(err, "more than one record in status ACCUMULATING")
 
 	// Halt the host zone and try again - it should fail
-	hostZone := s.MustGetHostZone()
+	hostZone, found := s.App.StakeibcKeeper.GetHostZone(s.Ctx, types.CelestiaChainId)
+	s.Require().True(found)
 	hostZone.Halted = true
-	s.App.StaketiaKeeper.SetHostZone(s.Ctx, hostZone)
+	s.App.StakeibcKeeper.SetHostZone(s.Ctx, hostZone)
 
 	err = s.App.StaketiaKeeper.PrepareUndelegation(s.Ctx, epochNumber)
-	s.Require().ErrorContains(err, "host zone is halted")
+	s.Require().ErrorContains(err, "host zone celestia is halted")
 }
 
 // ----------------------------------------------
@@ -464,7 +627,7 @@ func (s *KeeperTestSuite) SetupTestConfirmUndelegation(amountToUndelegate sdkmat
 	s.App.StaketiaKeeper.SetHostZone(s.Ctx, staketiaHostZone)
 
 	stakeibcHostZone := stakeibctypes.HostZone{
-		ChainId:          HostChainId,
+		ChainId:          types.CelestiaChainId,
 		RedemptionRate:   sdk.MustNewDecFromStr("1.1"),
 		TotalDelegations: delegatedBalance,
 	}
