+++ x/staketia/types/expected_keepers.go
@@ -49,4 +49,6 @@ type StakeibcKeeper interface {
 	GetHostZone(ctx sdk.Context, chainId string) (val stakeibctypes.HostZone, found bool)
 	GetActiveHostZone(ctx sdk.Context, chainId string) (hostZone stakeibctypes.HostZone, err error)
 	SetHostZone(ctx sdk.Context, hostZone stakeibctypes.HostZone)
+	RedeemStake(ctx sdk.Context, msg *stakeibctypes.MsgRedeemStake) (*stakeibctypes.MsgRedeemStakeResponse, error)
+	EnableRedemptions(ctx sdk.Context, chainId string) error
 }
